Move all reoutes into separate files and fix bugs with newly split DB helpers
